Projects/OCR-project/backend/requirements.txt
# Framework
fastapi==0.111.0
uvicorn[standard]==0.30.1
gunicorn==22.0.0

# Queue
celery[redis]==5.3.6
redis==5.0.4
flower==2.0.1

# DB
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1

# Validation / Auth
pydantic[email]==2.7.1
pydantic-settings==2.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# HTTP
httpx==0.27.0

# Storage
boto3==1.34.116
minio==7.2.7

# File validation
python-magic==0.4.27

# OCR - Digital PDF
pdfplumber==0.11.0
pypdf==4.2.0

# OCR - Scanned PDF
pdf2image==1.17.0
pytesseract==0.3.10
easyocr==1.7.1

# Image processing
Pillow==10.3.0
opencv-python-headless==4.9.0.80

# Language detection
langdetect==1.0.9

# Tokens
tiktoken==0.7.0
